refactor(CarService): domain폴더 하위로 이동
